Update ResultHome.tsx: reconstruct file
