+++ src/pages/resulthome/components/TypeDetail.tsx
@@ -1,0 +1,18 @@
+export const ResultDetail = () => {
+    return(
+
+        <div className="p-6">
+            <div className="font-['800'] mb-2">김푸앙님은 이런 생활 유형이에요!</div>
+            <div className="font-['700'] text-primary-gray mb-8">: 밤 12시에 취침하고, 오전 8시에 일어나요.매일 한 번의 방청소를 통해 깨끗한 환경을 유지하며,
+                    씻는 시간도 아침으로 일정하고 적절한 시간동안만 해요. 음식을 주로 방에서 섭취하지 않고, 비흡연자에요
+                    도서관에서 주로 공부하고, 노트북을 자주 사용해요. 룸메이트와 친하게 지내고 싶어해요. 잠버릇이 심해요.
+            </div>
+            <div className="font-['800'] mb-2">김푸앙님이 원하는 룸메이트 스타일은?</div>
+            <div className="font-['700'] text-primary-gray">: 밤 12시에 취침하고, 오전 8시에 일어났으면 좋겠어요.매일 한 번의 방청소를 통해 깨끗한 
+                   환경을 유지하길 바라며, 씻는 시간도 아침으로 일정하고 적절한 시간동안만 했으면 좋겠어요. 음식을 방에서 섭취하지 않았으면 좋겠고, 비흡연자였으면 좋겠어요
+                    도서관에서 주로 공부하고, 노트북 소음을 안 냈으면 좋겠어요.룸메이트도 저와 친하게 지냈으면 좋겠어요. 저는 잠귀가 어두워요
+            </div>
+        </div>
+            
+    )
+}
